Added db/models.base and updated models to be embedded
